JBPM-4651 - kie-spring jBPM - refactored and added integration tests
