--- kie-spring/src/test/java/org/kie/spring/jbpm/JTAEntityManagerSpringTest.java
@@ -1,103 +1,0 @@
-package org.kie.spring.jbpm;
-
-import java.util.List;
-import javax.naming.InitialContext;
-import javax.transaction.UserTransaction;
-
-import bitronix.tm.resource.jdbc.PoolingDataSource;
-import org.jbpm.process.audit.AuditLogService;
-import org.jbpm.process.audit.ProcessInstanceLog;
-import org.junit.After;
-import org.junit.Before;
-import org.junit.Test;
-import org.kie.api.runtime.KieSession;
-import org.kie.api.runtime.manager.RuntimeEngine;
-import org.kie.api.runtime.manager.RuntimeManager;
-import org.kie.api.runtime.process.ProcessInstance;
-import org.kie.api.task.TaskService;
-import org.kie.api.task.model.TaskSummary;
-import org.springframework.context.support.ClassPathXmlApplicationContext;
-
-import static junit.framework.Assert.*;
-
-public class JTAEntityManagerSpringTest extends AbstractJbpmSpringTest {
-
-    @Test
-    public void testSpringWithJTAAndEM() throws Exception{
-
-
-        context = new ClassPathXmlApplicationContext("jbpm/jta-em/singleton.xml");
-
-        RuntimeManager manager = (RuntimeManager) context.getBean("runtimeManager");
-
-        RuntimeEngine engine = manager.getRuntimeEngine(null);
-        KieSession ksession = engine.getKieSession();
-        TaskService taskService = engine.getTaskService();
-
-        ProcessInstance processInstance = ksession.startProcess("com.sample.bpmn.hello");
-
-        System.out.println("Process started");
-
-        AuditLogService logService = (AuditLogService) context.getBean("logService");
-        ProcessInstanceLog log = logService.findProcessInstance(processInstance.getId());
-        assertNotNull(log);
-
-        List<TaskSummary> tasks = taskService.getTasksAssignedAsPotentialOwner("john", "en-UK");
-        System.out.println("Found " + tasks.size() + " task(s) for user 'john'");
-        assertEquals(1, tasks.size());
-
-        long taskId = tasks.get(0).getId();
-        taskService.start(taskId, "john");
-        taskService.complete(taskId, "john", null);
-
-        tasks = taskService.getTasksAssignedAsPotentialOwner("mary", "en-UK");
-        System.out.println("Found " + tasks.size() + " task(s) for user 'mary'");
-        assertEquals(1, tasks.size());
-
-        taskId = tasks.get(0).getId();
-        taskService.start(taskId, "mary");
-        taskService.complete(taskId, "mary", null);
-
-        processInstance = ksession.getProcessInstance(processInstance.getId());
-        assertNull(processInstance);
-        System.out.println("Process instance completed");
-    }
-
-    @Test
-    public void testSpringWithJTAAndEMwithRollback() throws Exception{
-
-
-        context = new ClassPathXmlApplicationContext("jbpm/jta-em/singleton.xml");
-
-        RuntimeManager manager = (RuntimeManager) context.getBean("runtimeManager");
-
-        RuntimeEngine engine = manager.getRuntimeEngine(null);
-        KieSession ksession = engine.getKieSession();
-        TaskService taskService = engine.getTaskService();
-
-
-        AuditLogService logService = (AuditLogService) context.getBean("logService");
-
-        UserTransaction ut = (UserTransaction) new InitialContext().lookup("java:comp/UserTransaction");
-        ut.begin();
-        ProcessInstance processInstance = ksession.startProcess("com.sample.bpmn.hello");
-        long processInstanceId = processInstance.getId();
-        ut.rollback();
-
-        processInstance = ksession.getProcessInstance(processInstanceId);
-
-        if (processInstance == null) {
-            System.out.println("Process instance rolled back");
-        } else {
-            throw new IllegalArgumentException("Process instance not rolled back");
-        }
-
-        List<TaskSummary> tasks = taskService.getTasksAssignedAsPotentialOwner("john", "en-UK");
-        System.out.println("Found " + tasks.size() + " task(s) for user 'john'");
-        assertEquals(0, tasks.size());
-
-        ProcessInstanceLog log = logService.findProcessInstance(processInstanceId);
-        assertNull(log);
-    }
-
-}
