+++ kie-spring/src/test/java/org/kie/spring/jbpm/RuntimeManagerInitSpringTest.java
@@ -1,51 +1,69 @@
 package org.kie.spring.jbpm;
 
+import static org.junit.Assert.*;
+
+import java.util.Arrays;
+import java.util.Collection;
 import java.util.List;
 
-import javax.persistence.EntityManager;
-
-import bitronix.tm.resource.jdbc.PoolingDataSource;
 import org.jbpm.process.audit.AuditLogService;
 import org.jbpm.process.audit.ProcessInstanceLog;
-import org.junit.After;
-import org.junit.Before;
 import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.junit.runners.Parameterized;
 import org.kie.api.runtime.KieSession;
+import org.kie.api.runtime.manager.Context;
 import org.kie.api.runtime.manager.RuntimeEngine;
 import org.kie.api.runtime.manager.RuntimeManager;
 import org.kie.api.runtime.process.ProcessInstance;
 import org.kie.api.task.TaskService;
 import org.kie.api.task.model.TaskSummary;
+import org.kie.internal.runtime.manager.context.EmptyContext;
 import org.kie.internal.runtime.manager.context.ProcessInstanceIdContext;
 import org.springframework.context.support.ClassPathXmlApplicationContext;
 import org.springframework.transaction.TransactionStatus;
 import org.springframework.transaction.support.AbstractPlatformTransactionManager;
 import org.springframework.transaction.support.DefaultTransactionDefinition;
 
-import static junit.framework.Assert.*;
+/**
+ * Test used to verify correct initialization of runtime manager and transaction functionality in different configuration.
+ */
+@RunWith(Parameterized.class)
+public class RuntimeManagerInitSpringTest extends AbstractJbpmSpringTest {
 
-public class LocalEntityManagerSpringTest extends AbstractJbpmSpringTest {
+    @Parameterized.Parameters(name = "{index}: {0}")
+    public static Collection<Object[]> contextPath() {
+        Object[][] data = new Object[][] {
+                { "jbpm/jta-emf/singleton.xml", EmptyContext.get() },
+                { "jbpm/jta-em/singleton.xml", EmptyContext.get() },
+                { "jbpm/local-emf/singleton.xml", EmptyContext.get() },
+                { "jbpm/local-em/singleton.xml", EmptyContext.get() },
+                { "jbpm/jta-emf/per-process-instance.xml", ProcessInstanceIdContext.get() },
+                { "jbpm/jta-em/per-process-instance.xml", ProcessInstanceIdContext.get() },
+                { "jbpm/local-emf/per-process-instance.xml", ProcessInstanceIdContext.get() },
+                { "jbpm/local-em/per-process-instance.xml", ProcessInstanceIdContext.get() },
+                { "jbpm/jta-emf/per-request.xml", EmptyContext.get() },
+                { "jbpm/jta-em/per-request.xml", EmptyContext.get() },
+                { "jbpm/local-emf/per-request.xml", EmptyContext.get() },
+                { "jbpm/local-em/per-request.xml", EmptyContext.get() }
+        };
+        return Arrays.asList(data);
+    };
+
+    @Parameterized.Parameter(0)
+    public String contextPath;
+
+    @Parameterized.Parameter(1)
+    public Context<?> runtimeManagerContext;
 
     @Test
-    public void testSpringWithJTAAndEM() throws Exception{
+    public void testSimpleTaskInvocation() throws Exception{
 
+        context = new ClassPathXmlApplicationContext(contextPath);
 
-        context = new ClassPathXmlApplicationContext("jbpm/local-em/per-process-instance.xml");
-        EntityManager em = (EntityManager) context.getBean("jbpmEM");
-        // check that there is no sessions in db
-        List<?> sessions = em.createQuery("from SessionInfo").getResultList();
-        assertNotNull(sessions);
-        assertEquals(0, sessions.size());
-
-        AbstractPlatformTransactionManager aptm = (AbstractPlatformTransactionManager) context.getBean( "jbpmTxManager" );
         RuntimeManager manager = (RuntimeManager) context.getBean("runtimeManager");
 
-        // after creating per process instance manager init creates temp session that shall be directly destroyed
-        sessions = em.createQuery("from SessionInfo").getResultList();
-        assertNotNull(sessions);
-        assertEquals(0, sessions.size());
-
-        RuntimeEngine engine = manager.getRuntimeEngine(ProcessInstanceIdContext.get());
+        RuntimeEngine engine = manager.getRuntimeEngine(runtimeManagerContext);
         KieSession ksession = engine.getKieSession();
         TaskService taskService = engine.getTaskService();
 
@@ -54,10 +72,7 @@ public class LocalEntityManagerSpringTest extends AbstractJbpmSpringTest {
         System.out.println("Process started");
 
         AuditLogService logService = (AuditLogService) context.getBean("logService");
-        DefaultTransactionDefinition def = new DefaultTransactionDefinition();
-        TransactionStatus status = aptm.getTransaction(def);
         ProcessInstanceLog log = logService.findProcessInstance(processInstance.getId());
-        aptm.commit(status);
         assertNotNull(log);
 
         List<TaskSummary> tasks = taskService.getTasksAssignedAsPotentialOwner("john", "en-UK");
@@ -79,21 +94,21 @@ public class LocalEntityManagerSpringTest extends AbstractJbpmSpringTest {
         processInstance = ksession.getProcessInstance(processInstance.getId());
         assertNull(processInstance);
         System.out.println("Process instance completed");
+
+        manager.disposeRuntimeEngine(engine);
     }
 
     @Test
-    public void testSpringWithJTAAndEMwithRollback() throws Exception{
+    public void testSimpleTaskInvocationWithRollback() throws Exception{
 
-
-        context = new ClassPathXmlApplicationContext("jbpm/local-em/per-process-instance.xml");
+        context = new ClassPathXmlApplicationContext(contextPath);
 
         AbstractPlatformTransactionManager aptm = (AbstractPlatformTransactionManager) context.getBean( "jbpmTxManager" );
         RuntimeManager manager = (RuntimeManager) context.getBean("runtimeManager");
 
-        RuntimeEngine engine = manager.getRuntimeEngine(ProcessInstanceIdContext.get());
+        RuntimeEngine engine = manager.getRuntimeEngine(runtimeManagerContext);
         KieSession ksession = engine.getKieSession();
         TaskService taskService = engine.getTaskService();
-
 
         AuditLogService logService = (AuditLogService) context.getBean("logService");
 
@@ -115,10 +130,9 @@ public class LocalEntityManagerSpringTest extends AbstractJbpmSpringTest {
         System.out.println("Found " + tasks.size() + " task(s) for user 'john'");
         assertEquals(0, tasks.size());
 
-        def = new DefaultTransactionDefinition();
-        status = aptm.getTransaction(def);
         ProcessInstanceLog log = logService.findProcessInstance(processInstanceId);
-        aptm.commit(status);
         assertNull(log);
+
+        manager.disposeRuntimeEngine(engine);
     }
 }
