--- kie-spring/src/test/resources/jbpm/jpa-local-lane/jpa-spring.xml
@@ -1,26 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<beans xmlns="http://www.springframework.org/schema/beans"
-       xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-       xmlns:util="http://www.springframework.org/schema/util"
-       xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans-3.0.xsd
-                          http://www.springframework.org/schema/util http://www.springframework.org/schema/util/spring-util-3.0.xsd">
-
-  <bean id="jbpmEMF" class="org.springframework.orm.jpa.LocalContainerEntityManagerFactoryBean">
-    <property name="persistenceUnitName" value="org.jbpm.persistence.spring.local"/>
-    <property name="persistenceXmlLocation" value="classpath:jbpm/persistence-local.xml"/>
-  </bean>
-
-  <bean id="jbpmTxManager" class="org.springframework.orm.jpa.JpaTransactionManager">
-    <property name="entityManagerFactory" ref="jbpmEMF"/>
-  </bean>
-
-  <bean id="logService" class="org.jbpm.process.audit.JPAAuditLogService">
-    <constructor-arg>
-      <ref bean="jbpmEMF"/>
-    </constructor-arg>
-    <!-- <constructor-arg>
-      <value>STANDALONE_LOCAL_SPRING_SHARED_EM</value>
-    </constructor-arg> -->
-  </bean>
-
-</beans>
