+++ kie-spring/src/test/resources/jbpm/multiple-runtime-managers/local-emf-singleton.xml
@@ -7,19 +7,13 @@
 
   <import resource="classpath:jbpm/configuration-template/local-emf-spring.xml" />
 
-  <bean id="pessimisticLockingRuntimeEnvironment" parent="runtimeEnvironment">
-    <property name="pessimisticLocking" value="true" />
-    <property name="environmentEntries">
-      <map>
-        <entry key="TRANSACTION_LOCK_ENABLED">
-          <value type="java.lang.Boolean">false</value>
-        </entry>
-      </map>
-    </property>
-  </bean>
-
   <bean id="runtimeManager" class="org.kie.spring.factorybeans.RuntimeManagerFactoryBean" destroy-method="close">
     <property name="identifier" value="spring-rm"/>
-    <property name="runtimeEnvironment" ref="pessimisticLockingRuntimeEnvironment"/>
+    <property name="runtimeEnvironment" ref="runtimeEnvironment"/>
+  </bean>
+
+  <bean id="runtimeManagerTwo" class="org.kie.spring.factorybeans.RuntimeManagerFactoryBean" destroy-method="close">
+    <property name="identifier" value="spring-rm"/>
+    <property name="runtimeEnvironment" ref="runtimeEnvironment"/>
   </bean>
 </beans>
