--- kie-spring/src/test/resources/jbpm/pessimistic-lock/usergroups.properties
@@ -1,11 +1,0 @@
-admin=admins,managers,users,webdesigners,functionalanalysts
-krisv=admins,managers,users
-john=admins,managers,users,PM
-mary=admins,managers,users,HR
-sales-rep=admin,managers,users,sales
-frontDesk=users,PM
-doctor=users,PM
-manager=managers,HR
-salaboy=admins,users
-Administrator=Administrators
-ibek=admins,users
