+++ kie-spring/src/test/resources/jbpm/usergroup-callback/local-emf-singleton.xml
@@ -5,22 +5,18 @@
        xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans-3.0.xsd
                           http://www.springframework.org/schema/util http://www.springframework.org/schema/util/spring-util-3.0.xsd">
 
-  <import resource="classpath:jbpm/configuration-template/local-em-spring.xml" />
+  <import resource="classpath:jbpm/configuration-template/local-emf-spring.xml" />
 
-  <bean id="pessimisticLockingRuntimeEnvironment" parent="runtimeEnvironment">
-    <property name="pessimisticLocking" value="true" />
-    <property name="environmentEntries">
-      <map>
-        <entry key="TRANSACTION_LOCK_ENABLED">
-          <value type="java.lang.Boolean">false</value>
-        </entry>
-      </map>
-    </property>
+  <bean id="userGroupCallback" class="org.jbpm.services.task.identity.JBossUserGroupCallbackImpl">
+    <constructor-arg value="jbpm/usergroup-callback/usergroups.properties" />
+  </bean>
+
+  <bean id="userGroupCallbackRuntimeEnvironment" parent="runtimeEnvironment">
+    <property name="userGroupCallback" ref="userGroupCallback"/>
   </bean>
 
   <bean id="runtimeManager" class="org.kie.spring.factorybeans.RuntimeManagerFactoryBean" destroy-method="close">
     <property name="identifier" value="spring-rm"/>
-    <property name="runtimeEnvironment" ref="pessimisticLockingRuntimeEnvironment"/>
-    <property name="type" value="PER_REQUEST"/>
+    <property name="runtimeEnvironment" ref="userGroupCallbackRuntimeEnvironment"/>
   </bean>
 </beans>
