+++ kie-spring/src/test/resources/jbpm/usergroup-callback/usergroups.properties
@@ -1,0 +1,1 @@
+max=Standard
